+++ src/pages/Navbar/components/NavbarComponent.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Menu } from "semantic-ui-react";
 import { Link } from 'react-router-dom';
 
-import Profile from "./profileComponent";
+import Profile from "./ProfileComponent";
 
 
 const Navbar = ({authorised}) => {
